Reorganized solver and test runner
I wasn't happy with the bifurcated way testing was handled in the most recent version of my solver, and since I'm doing a lot of AoC outside of my usual pyCharm environment I wanted to add a few more tools and systems to support that.

Solvers now get instantiated, and if no RawData is passed in, it'll load that day's input data. a day should have a test_dayXX in the tests module, and when you want to run a day, it'll first run the tests and if the tests succeed, it'll proceed with actually solving. so now all testing (either on the known test data, or on smaller tests) happens in the unittests. This better supports the possibility of multiple test cases/answers within the puzzle itself.

This also eliminates the boilerplate if not data: data = self.processed step that was at the top of all the solver functions.